Single Server Queue
Calculates the amount of dropped packets based on buffer size, packet
arrival time and departure time, which is based on lambda (the
probability in which a packet arrives). Also calculates the overall
average of of packets in queue and the utilization ratio of the queue.

User inputs the amount of incoming packets and lambda